# GoMailFlows/docs/openapi.yaml
openapi: 3.1.0
info:
  title: HiTechCloud MailFlows API
  version: 1.0.0
  description: |
    HiTechCloud MailFlows - Enterprise Email SaaS Platform API
    
    ## Features
    - **Domain Management**: Register and manage email domains with automatic DNS provisioning
    - **Email Accounts**: Create and manage email mailboxes
    - **SMTP Configuration**: Per-domain or global SMTP relay configuration
    - **JMAP Support**: Full JMAP protocol (RFC 8620/8621) for modern email access
    - **Email Marketing**: Campaign management, subscriber lists, analytics
    - **SaaS Packages**: Flexible service plans with granular feature control
    - **Admin Panel**: Complete system administration capabilities
    - **DNS Templates**: Customizable DNS record templates for domain provisioning
    - **MCP Integration**: Model Context Protocol support for AI-powered management
    
    ## Authentication
    All authenticated endpoints require a Bearer token in the Authorization header.
    Obtain tokens via `/api/v1/auth/login` or `/api/v1/auth/register`.
    
    ## Rate Limiting
    API endpoints are rate-limited. Limits vary by endpoint and user package.
  contact:
    name: HiTechCloud Support
    url: https://hitechcloud.io/support
    email: support@hitechcloud.io
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.mailflows.hitechcloud.io/api/v1
...
security:
  - BearerAuth: []
paths:
...
  /marketing/campaigns:
    get:
      tags: [marketing]
      summary: List marketing campaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, scheduled, sending, sent, paused, cancelled]
      responses:
        '200':
          description: Campaign list

    post:
      tags: [marketing]
      summary: Create marketing campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
      responses:
        '201':
          description: Campaign created
  /marketing/campaigns/{id}/send:
    post:
      tags: [marketing]
      summary: Send campaign now
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Campaign queued
  /marketing/campaigns/{id}/stats:
    get:
      tags: [marketing]
      summary: Get campaign statistics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Campaign statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignStats'
  /marketing/lists:
    get:
      tags: [marketing]
      summary: List subscriber lists
      responses:
        '200':
          description: List of subscriber lists

    post:
      tags: [marketing]
      summary: Create subscriber list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListRequest'
      responses:
        '201':
          description: List created
  /marketing/subscribers/bulk:
    post:
      tags: [marketing]
      summary: Bulk add subscribers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkAddSubscribersRequest'
      responses:
        '200':
          description: Bulk add results

  # ============================================================
  # Admin
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login or /auth/register
  schemas:
...
    CreateCampaignRequest:
      type: object
      required: [name, subject, from_email]
      properties:
        name:
          type: string
        subject:
          type: string
        from_name:
          type: string
        from_email:
          type: string
          format: email
        reply_to:
          type: string
          format: email
        html_content:
          type: string
        plain_text:
          type: string
        tags:
          type: array
          items:
            type: string
    CampaignStats:
      type: object
      properties:
        campaign_id:
          type: string
          format: uuid
        total_recipients:
          type: integer
        sent:
          type: integer
        opens:
          type: integer
        clicks:
          type: integer
        bounces:
          type: integer
        unsubscribes:
          type: integer
        spam_reports:
          type: integer
        open_rate:
          type: number
        click_rate:
          type: number
        bounce_rate:
          type: number
    CreateListRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
    BulkAddSubscribersRequest:
      type: object
      required: [list_id, subscribers]
      properties:
        list_id:
          type: string
          format: uuid
        subscribers:
          type: array
          minItems: 1
          items:
            type: object
            required: [email]
            properties:
              email:
                type: string
                format: email
              first_name:
                type: string
              last_name:
                type: string
              meta_data:
                type: object
                additionalProperties:
                  type: string
...